Move fiat-units in constants directory
